pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-tg"
version = "0.2.0"
description = "Telegram bot for remote control of Claude Code / OpenAI Codex CLI sessions"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "sawierro" }]
license = { text = "MIT" }
dependencies = [
    "python-telegram-bot>=21.0,<22.0",
    "aiosqlite>=0.19.0,<1.0",
    "python-dotenv>=1.0.0,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0",
    "pytest-asyncio>=0.23.0,<1.0",
    "pytest-cov>=5.0.0,<7.0",
    "ruff>=0.6.0,<1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["bot*"]
exclude = ["tests*", "docs*", "scripts*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "integration: integration tests that may touch real subsystems",
]

[tool.coverage.run]
source = ["bot"]
omit = [
    "bot/__init__.py",
    "bot/claude_runner.py",
    "bot/external_sessions.py",
]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true
skip_covered = false
precision = 1

[tool.ruff]
line-length = 110
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # handled by formatter; we allow occasional long lines
    "B008",  # function calls in default arguments (used for field(default_factory=...))
    "SIM105", # try/except/pass is clearer than contextlib.suppress in many places
    "SIM108", # ternary preference — not worth the churn
    "SIM112", # Windows env vars (SystemRoot) are case-insensitive
    "SIM117", # combined with-statements — harms readability in some cases
    "B017",   # blind Exception in tests is fine for "should raise something" assertions
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E741"]
